增加查询接收数据的基础功能；修改了函数变量的文件位置；暂时约定了数据帧格式

+++ APP/include/General.h
@@ -9,7 +9,7 @@
 class ClassGeneral
 {
     public:
-    Uint16 CRC16Modbus(Uint8 *Addr, Uint16 Num);
+    Uint16 CRC16Modbus(Uint16 *Addr, Uint16 Num);
 
 };
