+++ APP/source/General.cpp
@@ -2,7 +2,7 @@
 
 ClassGeneral General;
 
-Uint16 ClassGeneral::CRC16Modbus(Uint8 *Addr, Uint16 Num)  
+Uint16 ClassGeneral::CRC16Modbus(Uint16 *Addr, Uint16 Num)  
 {
     Uint16 CRC = 0xFFFF;
     while (Num--)
